Seperate model testing into multiple files
